started working on Mapper interfaces

+++ src/main/java/rent.properly/maintenanceRequest/MaintenanceRequest.java
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 import rent.properly.property.Property;
 import rent.properly.tenant.Tenant;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
 
@@ -37,6 +38,6 @@ public class MaintenanceRequest {
     @Enumerated(EnumType.STRING)
     private MaintenanceRequestStatus status;
 
-    private BigInteger cost;
+    private BigDecimal cost;
 
 }
